requirements.txt
# Core Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# Deep Learning Framework
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Image Processing
Pillow>=8.3.0
opencv-python>=4.5.0

# Jupyter and Development Tools
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Additional Utilities
tqdm>=4.62.0
tensorboard>=2.8.0

# Optional: For better performance
numba>=0.56.0



